# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Test Doc",
        "description": "This is a simple API documentation.",
        "contact": {
            "email": "tugrulbo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://wwww.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1.2"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "List of all products",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many item to return at one time",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A page array of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpedted Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/products/": {
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "Create a product",
                "operationId": "store",
                "requestBody": {
                    "description": "Store a product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpedted Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{productId}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Info for a specific product",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the product to retrieve",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpedted Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "product"
                ],
                "summary": "Update a product",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the product to update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Update a product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpedted Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "product"
                ],
                "summary": "Delete a product",
                "operationId": "Delete",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The id column of the product to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpedted Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Product Model",
                "properties": {
                    "id": {
                        "description": "id column",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            }
        }
    }
}